[PR1a] Extract _as_bool to src/utils/helpers.py
Phase 1 refactoring: Move pure function _as_bool to dedicated utils module.

Changes:
- Create src/utils/helpers.py with _as_bool function
- Import _as_bool from utils module in main.py (re-export for backward compatibility)
- Remove original _as_bool function definition from main.py

Backward compatibility maintained:
- 'from src.main import _as_bool' still works (via re-export)
- Tests should patch via 'src.main._as_bool' (not 'src.utils.helpers._as_bool')

Verification:
- All 32 contract tests pass (import, patch, bootstrap)
- Route-map guard passes (9 tests)

See: docs/PHASE1_MAIN_PY_REFACTORING_PLAN.md
